App/requirements.txt
flask
mysql-connector-python
hvac